requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0

# LLM & Agents
langchain>=0.2.0
langchain-groq>=0.1.3
langchain-community>=0.2.0

# RAG / Vector Store
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

# OCR (Image Input)
easyocr>=1.7.1
Pillow>=10.0.0

# Audio Input (Speech to Text)
openai-whisper>=20231117

# Utilities
numpy>=1.24.0
